feat: add role and permission for bill item

+++ app/Http/Controllers/Admin/BillItemController.php
@@ -6,6 +6,7 @@ use App\Models\BillItem;
 use Illuminate\Http\Request;
 use Yajra\DataTables\DataTables;
 use App\Http\Controllers\Controller;
+use Illuminate\Support\Facades\Auth;
 use App\Http\Requests\Admin\BillItemRequest;
 
 class BillItemController extends Controller
@@ -15,15 +16,19 @@ class BillItemController extends Controller
      */
     public function index()
     {
+        if (!Auth::user()->can('Manage Item Bayar')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
+
         if (request()->ajax()) {
-            $data = BillItem::orderBy('name')->get();
+            $data = BillItem::orderBy('name');
             return DataTables::of($data)
                 ->addColumn('action', function ($data) {
                     $actionEdit = route('bill-item.edit', $data->id);
                     $actionDelete = route('bill-item.destroy', $data->id);
                     return "<div class='d-flex justify-content-center'>" .
-                        view('components.action.edit', ['action' => $actionEdit]) .
-                        view('components.action.delete', ['action' => $actionDelete, 'id' => $data->id]) .
+                        view('components.action.edit', ['action' => $actionEdit, 'name' => 'Item Bayar']) . '&nbsp;' .
+                        view('components.action.delete', ['action' => $actionDelete, 'id' => $data->id, 'name' => 'Item Bayar']) .
                         "</div>";
                 })
                 ->rawColumns(['action'])
@@ -37,6 +42,9 @@ class BillItemController extends Controller
      */
     public function create()
     {
+        if (!Auth::user()->can('Create Item Bayar')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         return view('admins.bill-item.create-edit');
     }
 
@@ -45,6 +53,9 @@ class BillItemController extends Controller
      */
     public function store(BillItemRequest $request)
     {
+        if (!Auth::user()->can('Create Item Bayar')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         BillItem::create($request->validated());
         return redirect()->route('bill-item.index')->with('success', 'Pos Bayar berhasil ditambahkan');
     }
@@ -62,6 +73,9 @@ class BillItemController extends Controller
      */
     public function edit(BillItem $billItem)
     {
+        if (!Auth::user()->can('Edit Item Bayar')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         return view('admins.bill-item.create-edit', compact('billItem'));
     }
 
@@ -70,6 +84,9 @@ class BillItemController extends Controller
      */
     public function update(BillItemRequest $request, BillItem $billItem)
     {
+        if (!Auth::user()->can('Edit Item Bayar')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         $billItem->update($request->validated());
         return redirect()->route('bill-item.index')->with('success', 'Pos Bayar berhasil diperbarui');
     }
@@ -79,6 +96,9 @@ class BillItemController extends Controller
      */
     public function destroy(BillItem $billItem)
     {
+        if (!Auth::user()->can('Delete Item Bayar')) {
+            return redirect()->back()->with('error', 'Maaf, Anda tidak memiliki akses untuk halaman tersebut');
+        }
         $billItem->delete();
         return redirect()->route('bill-item.index')->with('success', 'Pos Bayar berhasil dihapus');
     }
